Take out functions render statistic into view component to utils stats.js
