+++ src/utils/stats.js
@@ -1,146 +1,220 @@
 import Chart from 'chart.js';
 import '../../node_modules/flatpickr/dist/flatpickr.min.css';
 import ChartDataLabels from 'chartjs-plugin-datalabels';
+import { BAR_HEIGHT, POINT_TYPE_LIST } from '../consts.js';
+import { getTypesCount, getTypesMoney, getTypesTimeDuration } from '../utils/common.js';
+import { getDurationDate } from '../utils/date-helper.js';
 
-const moneyCtx = document.querySelector('#money');
-const typeCtx = document.querySelector('#type');
-const timeCtx = document.querySelector('#time');
+export const moneyChartRender = (moneyCtx, points) => {
+  moneyCtx.height = BAR_HEIGHT * points.length / 2;
+  const data = getTypesMoney(points, POINT_TYPE_LIST);
 
-// Рассчитаем высоту канваса в зависимости от того, сколько данных в него будет передаваться
-const BAR_HEIGHT = 55;
-moneyCtx.height = BAR_HEIGHT * 5;
-typeCtx.height = BAR_HEIGHT * 5;
-timeCtx.height = BAR_HEIGHT * 5;
-
-export const moneyChart = new Chart(moneyCtx, {
-  plugins: [ChartDataLabels],
-  type: 'horizontalBar',
-  data: {
-    labels: ['TAXI', 'BUS', 'TRAIN', 'SHIP', 'FLIGHT', 'DRIVE'],
-    datasets: [{
-      data: [400, 300, 200, 160, 150, 100],
-      backgroundColor: '#ffffff',
-      hoverBackgroundColor: '#ffffff',
-      anchor: 'start',
-      barThickness: 44,
-      minBarLength: 50,
-    }],
-  },
-  options: {
-    responsive: false,
-    plugins: {
-      datalabels: {
-        font: {
-          size: 13,
+  return new Chart(moneyCtx, {
+    plugins: [ChartDataLabels],
+    type: 'horizontalBar',
+    data: {
+      labels: POINT_TYPE_LIST,
+      datasets: [{
+        data,
+        backgroundColor: '#ffffff',
+        hoverBackgroundColor: '#ffffff',
+        anchor: 'start',
+        barThickness: 44,
+        minBarLength: 50,
+      }],
+    },
+    options: {
+      responsive: false,
+      plugins: {
+        datalabels: {
+          font: {
+            size: 13,
+          },
+          color: '#000000',
+          anchor: 'end',
+          align: 'start',
+          formatter: (val) => `€ ${val}`,
         },
-        color: '#000000',
-        anchor: 'end',
-        align: 'start',
-        formatter: (val) => `€ ${val}`,
+      },
+      title: {
+        display: true,
+        text: 'MONEY',
+        fontColor: '#000000',
+        fontSize: 23,
+        position: 'left',
+      },
+      scales: {
+        yAxes: [{
+          ticks: {
+            fontColor: '#000000',
+            padding: 5,
+            fontSize: 13,
+          },
+          gridLines: {
+            display: false,
+            drawBorder: false,
+          },
+        }],
+        xAxes: [{
+          ticks: {
+            display: false,
+            beginAtZero: true,
+          },
+          gridLines: {
+            display: false,
+            drawBorder: false,
+          },
+        }],
+      },
+      legend: {
+        display: false,
+      },
+      tooltips: {
+        enabled: false,
       },
     },
-    title: {
-      display: true,
-      text: 'MONEY',
-      fontColor: '#000000',
-      fontSize: 23,
-      position: 'left',
-    },
-    scales: {
-      yAxes: [{
-        ticks: {
-          fontColor: '#000000',
-          padding: 5,
-          fontSize: 13,
-        },
-        gridLines: {
-          display: false,
-          drawBorder: false,
-        },
-      }],
-      xAxes: [{
-        ticks: {
-          display: false,
-          beginAtZero: true,
-        },
-        gridLines: {
-          display: false,
-          drawBorder: false,
-        },
-      }],
-    },
-    legend: {
-      display: false,
-    },
-    tooltips: {
-      enabled: false,
-    },
-  },
-});
+  });
+};
 
-export const typeChart = new Chart(typeCtx, {
-  plugins: [ChartDataLabels],
-  type: 'horizontalBar',
-  data: {
-    labels: ['TAXI', 'BUS', 'TRAIN', 'SHIP', 'FLIGHT', 'DRIVE'],
-    datasets: [{
-      data: [4, 3, 2, 1, 1, 1],
-      backgroundColor: '#ffffff',
-      hoverBackgroundColor: '#ffffff',
-      anchor: 'start',
-      barThickness: 44,
-      minBarLength: 50,
-    }],
-  },
-  options: {
-    responsive: false,
-    plugins: {
-      datalabels: {
-        font: {
-          size: 13,
+export const typeChartRender = (typeCtx, points) => {
+  typeCtx.height = BAR_HEIGHT * points.length / 2;
+  const data = getTypesCount(points, POINT_TYPE_LIST);
+
+  return new Chart(typeCtx, {
+    plugins: [ChartDataLabels],
+    type: 'horizontalBar',
+    data: {
+      labels: POINT_TYPE_LIST,
+      datasets: [{
+        data,
+        backgroundColor: '#ffffff',
+        hoverBackgroundColor: '#ffffff',
+        anchor: 'start',
+        barThickness: 44,
+        minBarLength: 50,
+      }],
+    },
+    options: {
+      responsive: false,
+      plugins: {
+        datalabels: {
+          font: {
+            size: 13,
+          },
+          color: '#000000',
+          anchor: 'end',
+          align: 'start',
+          formatter: (val) => `${val}x`,
         },
-        color: '#000000',
-        anchor: 'end',
-        align: 'start',
-        formatter: (val) => `${val}x`,
+      },
+      title: {
+        display: true,
+        text: 'TYPE',
+        fontColor: '#000000',
+        fontSize: 23,
+        position: 'left',
+      },
+      scales: {
+        yAxes: [{
+          ticks: {
+            fontColor: '#000000',
+            padding: 5,
+            fontSize: 13,
+          },
+          gridLines: {
+            display: false,
+            drawBorder: false,
+          },
+        }],
+        xAxes: [{
+          ticks: {
+            display: false,
+            beginAtZero: true,
+          },
+          gridLines: {
+            display: false,
+            drawBorder: false,
+          },
+        }],
+      },
+      legend: {
+        display: false,
+      },
+      tooltips: {
+        enabled: false,
       },
     },
-    title: {
-      display: true,
-      text: 'TYPE',
-      fontColor: '#000000',
-      fontSize: 23,
-      position: 'left',
-    },
-    scales: {
-      yAxes: [{
-        ticks: {
-          fontColor: '#000000',
-          padding: 5,
-          fontSize: 13,
-        },
-        gridLines: {
-          display: false,
-          drawBorder: false,
-        },
-      }],
-      xAxes: [{
-        ticks: {
-          display: false,
-          beginAtZero: true,
-        },
-        gridLines: {
-          display: false,
-          drawBorder: false,
-        },
-      }],
-    },
-    legend: {
-      display: false,
-    },
-    tooltips: {
-      enabled: false,
-    },
-  },
-});
+  });
+};
 
+
+export const timeChartRender = (typeCtx, points) => {
+  typeCtx.height = BAR_HEIGHT * points.length / 2;
+  const data = getTypesTimeDuration(points, POINT_TYPE_LIST);
+
+  return new Chart(typeCtx, {
+    plugins: [ChartDataLabels],
+    type: 'horizontalBar',
+    data: {
+      labels: POINT_TYPE_LIST,
+      datasets: [{
+        data,
+        backgroundColor: '#ffffff',
+        hoverBackgroundColor: '#ffffff',
+        anchor: 'start',
+        barThickness: 44,
+        minBarLength: 50,
+      }],
+    },
+    options: {
+      responsive: false,
+      plugins: {
+        datalabels: {
+          font: {
+            size: 13,
+          },
+          color: '#000000',
+          anchor: 'end',
+          align: 'start',
+          formatter: (val) => `${getDurationDate(val)}`,
+        },
+      },
+      title: {
+        display: true,
+        text: 'TIME',
+        fontColor: '#000000',
+        fontSize: 23,
+        position: 'left',
+      },
+      scales: {
+        yAxes: [{
+          ticks: {
+            fontColor: '#000000',
+            padding: 5,
+            fontSize: 13,
+          },
+          gridLines: {
+            display: false,
+            drawBorder: false,
+          },
+        }],
+        xAxes: [{
+          ticks: {
+            display: false,
+            beginAtZero: true,
+          },
+          gridLines: {
+            display: false,
+            drawBorder: false,
+          },
+        }],
+      },
+      legend: {
+        display: false,
+      },
+      tooltips: {
+        enabled: false,
+      },
+    },
+  });
+};
