+++ src/view/site-statistics-view.js
@@ -1,226 +1,5 @@
+import { moneyChartRender, timeChartRender, typeChartRender } from '../utils/stats.js';
 import SiteSmartView from './site-smart-view.js';
-import Chart from 'chart.js';
-import '../../node_modules/flatpickr/dist/flatpickr.min.css';
-import ChartDataLabels from 'chartjs-plugin-datalabels';
-import { POINT_TYPE_LIST } from '../consts.js';
-import { getTypesCount, getTypesMoney, getTypesTimeDuration } from '../utils/common.js';
-import { getDurationDate } from '../utils/date-helper.js';
-
-const BAR_HEIGHT = 55;
-
-export const moneyChartRender = (moneyCtx, points) => {
-  moneyCtx.height = BAR_HEIGHT * points.length / 2;
-  const data = getTypesMoney(points, POINT_TYPE_LIST);
-
-  return new Chart(moneyCtx, {
-    plugins: [ChartDataLabels],
-    type: 'horizontalBar',
-    data: {
-      labels: POINT_TYPE_LIST,
-      datasets: [{
-        data,
-        backgroundColor: '#ffffff',
-        hoverBackgroundColor: '#ffffff',
-        anchor: 'start',
-        barThickness: 44,
-        minBarLength: 50,
-      }],
-    },
-    options: {
-      responsive: false,
-      plugins: {
-        datalabels: {
-          font: {
-            size: 13,
-          },
-          color: '#000000',
-          anchor: 'end',
-          align: 'start',
-          formatter: (val) => `€ ${val}`,
-        },
-      },
-      title: {
-        display: true,
-        text: 'MONEY',
-        fontColor: '#000000',
-        fontSize: 23,
-        position: 'left',
-      },
-      scales: {
-        yAxes: [{
-          ticks: {
-            fontColor: '#000000',
-            padding: 5,
-            fontSize: 13,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-        xAxes: [{
-          ticks: {
-            display: false,
-            beginAtZero: true,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-      },
-      legend: {
-        display: false,
-      },
-      tooltips: {
-        enabled: false,
-      },
-    },
-  });
-};
-
-export const typeChartRender = (typeCtx, points) => {
-  typeCtx.height = BAR_HEIGHT * points.length / 2;
-  const data = getTypesCount(points, POINT_TYPE_LIST);
-
-  return new Chart(typeCtx, {
-    plugins: [ChartDataLabels],
-    type: 'horizontalBar',
-    data: {
-      labels: POINT_TYPE_LIST,
-      datasets: [{
-        data,
-        backgroundColor: '#ffffff',
-        hoverBackgroundColor: '#ffffff',
-        anchor: 'start',
-        barThickness: 44,
-        minBarLength: 50,
-      }],
-    },
-    options: {
-      responsive: false,
-      plugins: {
-        datalabels: {
-          font: {
-            size: 13,
-          },
-          color: '#000000',
-          anchor: 'end',
-          align: 'start',
-          formatter: (val) => `${val}x`,
-        },
-      },
-      title: {
-        display: true,
-        text: 'TYPE',
-        fontColor: '#000000',
-        fontSize: 23,
-        position: 'left',
-      },
-      scales: {
-        yAxes: [{
-          ticks: {
-            fontColor: '#000000',
-            padding: 5,
-            fontSize: 13,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-        xAxes: [{
-          ticks: {
-            display: false,
-            beginAtZero: true,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-      },
-      legend: {
-        display: false,
-      },
-      tooltips: {
-        enabled: false,
-      },
-    },
-  });
-};
-
-
-export const timeChartRender = (typeCtx, points) => {
-  typeCtx.height = BAR_HEIGHT * points.length / 2;
-  const data = getTypesTimeDuration(points, POINT_TYPE_LIST);
-
-  return new Chart(typeCtx, {
-    plugins: [ChartDataLabels],
-    type: 'horizontalBar',
-    data: {
-      labels: POINT_TYPE_LIST,
-      datasets: [{
-        data,
-        backgroundColor: '#ffffff',
-        hoverBackgroundColor: '#ffffff',
-        anchor: 'start',
-        barThickness: 44,
-        minBarLength: 50,
-      }],
-    },
-    options: {
-      responsive: false,
-      plugins: {
-        datalabels: {
-          font: {
-            size: 13,
-          },
-          color: '#000000',
-          anchor: 'end',
-          align: 'start',
-          formatter: (val) => `${getDurationDate(val)}`,
-        },
-      },
-      title: {
-        display: true,
-        text: 'TIME',
-        fontColor: '#000000',
-        fontSize: 23,
-        position: 'left',
-      },
-      scales: {
-        yAxes: [{
-          ticks: {
-            fontColor: '#000000',
-            padding: 5,
-            fontSize: 13,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-        xAxes: [{
-          ticks: {
-            display: false,
-            beginAtZero: true,
-          },
-          gridLines: {
-            display: false,
-            drawBorder: false,
-          },
-        }],
-      },
-      legend: {
-        display: false,
-      },
-      tooltips: {
-        enabled: false,
-      },
-    },
-  });
-};
 
 const createStatisticsTemplate = () => (
   `<section class="statistics">
